small refactor into base trait
